requirements.txt
adjustText>=0.7.3
boltons>=20.2.1
ctxcore>=0.1.1
frozendict
matplotlib
multiprocessing
numba>=0.50.1
numpy>=1.18.5
pandas>=1.0.0
pyarrow>=0.11.1
pyscenic>=0.10.4
scanpy>=1.6.0
scipy
seaborn>=0.11.1
statsmodel>=0.11.1
tqdm
UpSetPlot>=0.4.1